+++ src/fire.class.cpp
@@ -2,4 +2,4 @@
 
 Fire::~Fire( void ) {}
 
-Fire::Fire( int x, int y, int status ) : Entity( FIRE, 0, x, y, status ) {}
+Fire::Fire( float x, float y, int status ) : Entity( FIRE, 0, x, y, status ) {}
