+++ src/wall.class.cpp
@@ -2,4 +2,4 @@
 
 Wall::~Wall( void ) {}
 
-Wall::Wall( int x, int y, int status ) : Entity( WALL, 0, x, y, status ) {}
+Wall::Wall( float x, float y, int status ) : Entity( WALL, 0, x, y, status ) {}
